Implementa testes para modificacao no registra-atendimento

+++ src/main/java/com/br/fiap/controller/VendedorController.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,10 +30,13 @@ public class VendedorController {
 	private FilaAtendimentoRepository filaAtendimentoRepository;
 	
 	@PostMapping("/registra-atendimento")
-	public ResponseEntity<Void> registraAtendimento(@RequestBody @Valid AtendimentoDTO formulario) {
-		this.clienteRepository.save(formulario.cliente().converteParaCliente());
+	public ResponseEntity<Void> registraAtendimento(@RequestBody @Valid AtendimentoDTO formulario,
+			@RequestHeader LocalAcessadoEnum localAcessado) {
+		if(localAcessado.equals(LocalAcessadoEnum.SITE)) {
+			this.clienteRepository.save(formulario.cliente().converteParaCliente());
+			this.filaAtendimentoRepository.deleteById(formulario.pegaIdFilaAtendimento());
+		}
 		this.atendimentoRepository.save(formulario.converteParaAtendimento());
-		this.filaAtendimentoRepository.deleteById(formulario.pegaIdFilaAtendimento());
 		return ResponseEntity
 				.status(HttpStatus.CREATED)
 				.build();
